+++ camel/Extraction/Metadata.php
@@ -7,6 +7,7 @@ use Knuckles\Camel\BaseDTO;
 class Metadata extends BaseDTO
 {
     public ?string $groupName;
+    public ?string $subgroup;
 
     /**
      * Name of the group that this group should be placed just before.